The delete/move operations now operate over entire graphemes
Summary: When you move/delete forward/backward your expectation is that you will be operating over one 'character'. To you, that usually means 'one grapheme'. For instance, an Emoji with a skin-tone modifier is made up of multiple Unicode codepoints, but you think of it as one glyph. When you press delete, you expect the entire thing to be deleted – not just the modifier.

Fixes #17.